Extract AI comment UI into LinkAiExtract component
Add LinkAiExtract.vue to encapsulate the AI-driven comment extraction UI and logic (fetch recent comments, select/edit comment, call AI extract API, emit result). Refactor LinkForm.vue to use the new component for new links (isNew computed) and remove duplicated AI extraction code and unused imports; rename the internal handler to applyAiExtractedResult and wire it to the component's @extract event. This centralizes AI extraction behavior and simplifies the link form.
