+++ console/src/components/LinkForm.vue
@@ -1,12 +1,10 @@
 <script lang="ts" setup>
-import { linkAiApiClient, linksConsoleApiClient } from "@/api";
-import type { LinkCommentAnalysisResult, LinkCommentDTO } from "@/api/generated";
+import { linksConsoleApiClient } from "@/api";
+import type { LinkCommentAnalysisResult } from "@/api/generated";
 import type { LinkFormState } from "@/types";
-import { IconClose, Toast, VButton, VLoading } from "@halo-dev/components";
-import { utils } from "@halo-dev/ui-shared";
-import { nextTick, onMounted, ref, shallowRef, toRaw } from "vue";
-import MdiCommentTextOutline from "~icons/mdi/comment-text-outline";
-import MdiRobot from "~icons/mdi/robot";
+import { Toast, VButton, VLoading } from "@halo-dev/components";
+import { computed, nextTick, onMounted, ref, shallowRef, toRaw } from "vue";
+import LinkAiExtract from "./LinkAiExtract.vue";
 import MdiWebRefresh from "~icons/mdi/web-refresh";
 import Rss2FillIcon from "~icons/mingcute/rss-2-fill";
 
@@ -117,56 +115,9 @@ const handleDiscoverFeed = async () => {
 
 const annotationsForm = ref();
 
-// AI comment extraction
-const showAiExtract = shallowRef(false);
-const isLoadingComments = shallowRef(false);
-const isExtracting = shallowRef(false);
-const recentComments = ref<LinkCommentDTO[]>([]);
-const selectedCommentName = shallowRef<string | undefined>(undefined);
-const manualCommentText = shallowRef("");
+const isNew = computed(() => !props.formState);
 
-async function handleFetchComments() {
-  if (isLoadingComments.value) return;
-  isLoadingComments.value = true;
-  try {
-    const { data } = await linkAiApiClient.ai.listRecentComments();
-    recentComments.value = data;
-    if (!data.length) {
-      Toast.info("暂无已审核的评论");
-    }
-  } catch {
-    // Halo's API interceptor shows request failure toasts.
-  } finally {
-    isLoadingComments.value = false;
-  }
-}
-
-function selectComment(comment: LinkCommentDTO) {
-  selectedCommentName.value = comment.name;
-  manualCommentText.value = comment.raw || comment.content || "";
-}
-
-async function handleAiExtract() {
-  const content = manualCommentText.value.trim();
-  if (!content) {
-    Toast.error("请选择一条评论或输入评论内容");
-    return;
-  }
-  if (isExtracting.value) return;
-  isExtracting.value = true;
-  try {
-    const { data: result } = await linkAiApiClient.ai.extractLinkFromComment({
-      body: content,
-    });
-    applyExtractedResult(result);
-    Toast.success("AI 识别成功");
-    showAiExtract.value = false;
-  } finally {
-    isExtracting.value = false;
-  }
-}
-
-function applyExtractedResult(result: LinkCommentAnalysisResult) {
+function applyAiExtractedResult(result: LinkCommentAnalysisResult) {
   if (result.url) data.value.url = result.url;
   if (result.displayName) data.value.displayName = result.displayName;
   if (result.logo) data.value.logo = result.logo;
@@ -241,114 +192,22 @@ async function onSubmit() {
 <template>
   <FormKit id="link-form" name="link-form" type="form" :config="{ validationVisibility: 'submit' }" @submit="onSubmit">
     <div>
-      <div class=":uno: md:grid md:grid-cols-4 md:gap-6">
-        <div class=":uno: md:col-span-1">
-          <div class=":uno: sticky top-0">
-            <span class=":uno: text-base text-gray-900 font-medium"> AI 识别 </span>
+      <template v-if="isNew">
+        <div class=":uno: md:grid md:grid-cols-4 md:gap-6">
+          <div class=":uno: md:col-span-1">
+            <div class=":uno: sticky top-0">
+              <span class=":uno: text-base text-gray-900 font-medium"> AI 识别 </span>
+            </div>
+          </div>
+          <div class=":uno: mt-5 md:col-span-3 md:mt-0">
+            <LinkAiExtract @extract="applyAiExtractedResult" />
           </div>
         </div>
-        <div class=":uno: mt-5 md:col-span-3 md:mt-0">
-          <div v-if="!showAiExtract" class=":uno: flex items-center gap-2">
-            <VButton size="sm" type="secondary" @click="showAiExtract = true">
-              <template #icon>
-                <MdiRobot class=":uno: size-4" />
-              </template>
-              从评论识别
-            </VButton>
-            <span class=":uno: text-xs text-gray-500">通过 AI 从访客评论中提取友链信息</span>
-          </div>
-          <div v-else class=":uno: space-y-4 rounded-xl border border-gray-200 bg-white p-4 shadow-sm">
-            <div class=":uno: flex items-center justify-between">
-              <div class=":uno: flex items-center gap-2">
-                <MdiRobot class=":uno: size-5 text-blue-600" />
-                <span class=":uno: text-sm font-semibold text-gray-900">从评论识别友链</span>
-              </div>
-              <button
-                type="button"
-                class=":uno: flex items-center justify-center rounded p-1 text-gray-400 transition-colors hover:bg-gray-100 hover:text-gray-600"
-                @click="showAiExtract = false"
-              >
-                <IconClose class=":uno: size-4" />
-              </button>
-            </div>
 
-            <p class=":uno: text-xs text-gray-500">从已审核的评论中提取网站地址、名称、Logo、描述和 RSS 订阅地址。</p>
-
-            <VButton size="sm" :loading="isLoadingComments" @click="handleFetchComments">
-              <template #icon>
-                <MdiCommentTextOutline class=":uno: size-4" />
-              </template>
-              获取最新评论
-            </VButton>
-
-            <div v-if="isLoadingComments" class=":uno: flex justify-center py-6">
-              <VLoading class=":uno: size-5 text-gray-400" />
-            </div>
-
-            <div
-              v-else-if="recentComments.length"
-              class=":uno: max-h-52 overflow-y-auto rounded-lg border border-gray-200 divide-y divide-gray-100"
-            >
-              <div
-                v-for="comment in recentComments"
-                :key="comment.name"
-                class=":uno: cursor-pointer px-3 py-2.5 transition-colors"
-                :class="selectedCommentName === comment.name ? 'bg-blue-50/60' : 'hover:bg-gray-50'"
-                @click="selectComment(comment)"
-              >
-                <div class=":uno: flex items-center justify-between gap-2">
-                  <div class=":uno: flex items-center gap-2 min-w-0">
-                    <span
-                      class=":uno: inline-block size-2 flex-none rounded-full"
-                      :class="selectedCommentName === comment.name ? 'bg-blue-500' : 'bg-gray-300'"
-                    ></span>
-                    <span class=":uno: text-sm font-medium text-gray-900 truncate">
-                      {{ comment.ownerName || "匿名" }}
-                    </span>
-                  </div>
-                  <span class=":uno: text-xs text-gray-400 flex-none">
-                    {{ utils.date.timeAgo(comment.creationTime) }}
-                  </span>
-                </div>
-                <p class=":uno: mt-1 text-xs text-gray-500 line-clamp-2 pl-4">
-                  {{ comment.raw || comment.content }}
-                </p>
-              </div>
-            </div>
-
-            <div v-else class=":uno: rounded-lg border border-dashed border-gray-200 py-5 text-center">
-              <p class=":uno: text-sm text-gray-400">暂无已审核的评论</p>
-              <p class=":uno: mt-1 text-xs text-gray-400">可手动在下方粘贴评论内容</p>
-            </div>
-
-            <div>
-              <label class=":uno: block text-xs font-medium text-gray-700 mb-1.5">
-                评论内容
-                <span class=":uno: font-normal text-gray-400">（可手动粘贴或编辑）</span>
-              </label>
-              <textarea
-                v-model="manualCommentText"
-                rows="3"
-                class=":uno: w-full rounded-lg border border-gray-200 px-3 py-2 text-sm transition-colors focus:border-blue-500 focus:outline-none focus:ring-2 focus:ring-blue-500/20"
-                placeholder="选择上方评论或手动粘贴内容，AI 将从中提取友链信息..."
-              ></textarea>
-            </div>
-
-            <div class=":uno: flex justify-end">
-              <VButton size="sm" type="secondary" :loading="isExtracting" @click="handleAiExtract">
-                <template #icon>
-                  <MdiRobot class=":uno: size-4" />
-                </template>
-                AI 识别
-              </VButton>
-            </div>
-          </div>
+        <div class=":uno: py-5">
+          <div class=":uno: border-t border-gray-200"></div>
         </div>
-      </div>
-
-      <div class=":uno: py-5">
-        <div class=":uno: border-t border-gray-200"></div>
-      </div>
+      </template>
 
       <div class=":uno: md:grid md:grid-cols-4 md:gap-6">
         <div class=":uno: md:col-span-1">
